rosa/requirements.txt
av
gpiozero
luma.oled
pillow
psutil
requests
RPi.GPIO
lgpio
